% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmdist.R
\name{gmmdist}
\alias{gmmdist}
\title{Distance Measures between Multisets using Gaussian Mixture Models}
\usage{
gmmdist(datalist, maxk = 5, method = c("L2"), as.dist = FALSE)
}
\description{
Distance Measures between Multisets using Gaussian Mixture Models
}
