% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_hiddenR.R
\name{hidden_geigen}
\alias{hidden_geigen}
\title{It mimics the behavior of 'geigen' function with normalization added}
\usage{
hidden_geigen(A, B, normalize = TRUE)
}
\description{
It mimics the behavior of 'geigen' function with normalization added
}
\keyword{internal}
