% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-maotai.R
\docType{package}
\name{maotai}
\alias{maotai}
\alias{maotai-package}
\title{Tools for Matrix Algebra, Optimization and Inference Problems}
\description{
Tools for Matrix Algebra, Optimization and Inference Problems
}
