% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-maotai.R
\docType{package}
\name{package-maotai}
\alias{package-maotai}
\title{Tools for Matrix Algebra, Optimization and Inference}
\description{
Matrix is an universal and sometimes primary object/unit in applied mathematics and statistics. We provide a number of algorithms for selected problems in optimization and statistical inference.
}
