% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pdeterminant.R
\name{pdeterminant}
\alias{pdeterminant}
\title{Pseudo-Determinant
# (1) Holbrook's Note on Pseudo-Determinant -------------------------------}
\usage{
pdeterminant(A)
}
\description{
Pseudo-Determinant
# (1) Holbrook's Note on Pseudo-Determinant -------------------------------
}
