% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.R
\name{trio}
\alias{trio}
\title{Trace Ratio Optimation}
\usage{
trio(A, B, C, dim = 2, method = c("2003Guo", "2007Wang", "2009Jia",
  "2012Ngo"), maxiter = 1000, eps = 1e-10)
}
\description{
Trace Ratio Optimation
}
